+++ rustfst/src/algorithms/fst_convert.rs
@@ -26,13 +26,7 @@ where
         }
     }
 
-    if let Some(isymt) = ifst.input_symbols() {
-        ofst.set_input_symbols(isymt);
-    }
-
-    if let Some(osymt) = ifst.output_symbols() {
-        ofst.set_output_symbols(osymt);
-    }
+    ofst.set_symts_from_fst(ifst);
 
     ofst
 }